W0-2: widen EffectKind byte -> ushort (behaviour-free, pin-neutral)
Six of the eight bits a byte offers were already spent (SetDisposition ..
EndConversation). Two economy effects and it is full; widening after the
parallel economy lanes start queueing effects becomes a cross-lane refactor
touching CapabilityComputer, EffectValidator, the applied-effect event and the
Sim.Llm tool-schema builders. So it happens now, alone, width only — no new
effect kinds, no behaviour change.

Nothing else had to move with it. The audit, for the record:

- NOT persisted. EffectKind appears in exactly four sim files
  (CitizenEffects.cs, CapabilityComputer.cs, CitizenMemory.cs:305 inside Tick,
  and the Sim.Llm DTO layer). MemorySystem.CaptureState / RestoreState /
  StateChecksum never touch it, no SaveWriter chapter writes it, and no
  IStatefulSystem folds it. No version branch, no truncation guard needed.
- NOT hashed. The determinism pin 26907c23d7e48a5c is unmoved and the twin
  hashes MATCH, as ECONOMY-PLAN.md 3.1 budgets for this row.
- NOT on the wire. WireFormat.ChatEffect sends an already-rendered human note
  string; the client's chat entry is {kind:'effect', text}. No numeric width
  crosses to client/.
- NOT in the tool schemas. Both propose_effect builders (PromptBuilder's frozen
  block and AnthropicBackend.WriteProposeEffectTool) bound the enum by NAME.
- No (byte) cast of an EffectKind exists anywhere in the tree; the only byte
  casts near effects are TileFlags/RelationType/HistoryKind, untouched.
- Perilune.Llm.EffectKind (IChatBackend.cs) is a separate int-backed proposal
  DTO with no width pressure; deliberately left alone.

EffectKindWidthTests drives the two paths that actually carry the value rather
than asserting sizeof: the event bus into MemorySystem's promise rule (a kind of
1<<8|AgreeTask must NOT form a promise memory - under a byte store it truncates
to exactly AgreeTask and forges one), and the CapabilityManifest legal-set into
TurnPlan (a top-bit grant must survive the copy). Both mutations named in the
doc comments were applied and observed red, then reverted.

611 dotnet + 207 node green via ./ci.sh in-worktree (607 -> 611, +4 new).

+++ sim/Sim.Core/Effects/CitizenEffects.cs
@@ -70,9 +70,24 @@ namespace Perilune.Sim
     /// <summary>
     /// Effect discriminator. Flags so <see cref="CapabilityManifest.LegalEffects"/>
     /// can express a legal-set; single values identify one effect in events.
+    ///
+    /// WIDTH — <c>ushort</c>, deliberately (ECONOMY-PLAN.md §0 W0-2). Six of the eight
+    /// bits a <c>byte</c> offers were already spent, so the economy's first two effects
+    /// would have filled it; widening later would have been a cross-lane refactor. The
+    /// widening is behaviour-free and pin-neutral: <see cref="EffectKind"/> is not
+    /// written to any save chapter, not folded into any <c>StateChecksum</c>, and not
+    /// carried numerically on the wire (the web skin sends an already-rendered effect
+    /// NOTE string — <c>WireFormat.ChatEffect</c>), and the provider tool schemas bound
+    /// the enum by NAME, not ordinal. Nothing may narrow it back on the way through:
+    /// there is no <c>(byte)</c> cast of an EffectKind anywhere in the tree, and
+    /// <c>EffectKindWidthTests</c> pins that flags above bit 7 survive the two paths
+    /// that actually carry the value (the event bus into MemorySystem's rule table, and
+    /// the CapabilityManifest legal-set into TurnPlan).
+    ///
+    /// Rows stay append-only and stay powers of two; the next free bit is 1 &lt;&lt; 6.
     /// </summary>
     [Flags]
-    public enum EffectKind : byte
+    public enum EffectKind : ushort
     {
         None = 0,
         SetDisposition = 1 << 0,
